chore: encapsulate into EmbedDashboardParameters
